Support prevent_default for key and mouse event
